Save remaining lesson files

+++ lesson-01/exercise3.py
@@ -1,4 +1,4 @@
-number = float(input("Enter a number: "))
+number = int(input("Enter a number: "))
 result = number * 10
 
 print(f"{number} times 10 is {result}")
